Fixed config issues
