+++ client/flutter/fadzmaq/.vscode/launch.json
@@ -1,0 +1,24 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+
+
+    // Make sure you are running VS code from this root folder to build correctly
+
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Local",
+            "request": "launch",
+            "type": "dart",
+            "program": "lib/main.dart",
+        },
+        {
+            "name": "Test Server",
+            "request": "launch",
+            "type": "dart",
+            "program": "lib/main_testserver.dart",
+        }
+    ]
+}
