--- client/flutter/fadzmaq/lib/main_local.dart
@@ -1,15 +1,0 @@
-import 'package:fadzmaq/app_config.dart';
-import 'package:fadzmaq/app.dart';
-import 'package:fadzmaq/resource/config_local.dart';
-
-import 'package:flutter/material.dart';
-
-
-void main() {
-  var configuredApp = AppConfig(
-    appConfig: ConfigResourceDev(),
-    child: App(),
-  );
-
-  runApp(configuredApp);
-}
