+++ client/flutter/fadzmaq/lib/main_testserver.dart
@@ -4,7 +4,7 @@ import 'package:fadzmaq/resource/config_testserver.dart';
 
 import 'package:flutter/material.dart';
 
-
+// entry point to load default config for test server
 void main() {
   var configuredApp = AppConfig(
     appConfig: ConfigResourceDefault(),
