+++ client/flutter/fadzmaq/lib/resource/config_local.dart
@@ -2,5 +2,5 @@ import 'package:fadzmaq/app_config.dart';
 
 class ConfigResourceDev implements ConfigResource {
   @override
-  String testString = "Test string dev";
+  String testString = "Test string devopment";
 }
