+++ client/flutter/fadzmaq/lib/resource/config_testserver.dart
@@ -2,5 +2,5 @@ import 'package:fadzmaq/app_config.dart';
 
 class ConfigResourceDefault implements ConfigResource {
   @override
-  String testString = "Test string default";
+  String testString = "Test string test server";
 }
